+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+1.12.4 / 2014-08-15
+===================
+
+* Internal test cleanup
+* Extract bintray functionality to [nebula-bintray-plugin](https://github.com/nebula-plugins/nebula-bintray-plugin)
+
+1.12.3 / 2014-07-02
+===================
+
+* Fix maven publication name
+* nebula-project-plugin to 1.12.7
+
 1.12.2 / 2014-06-11
 ===================